added backdrops to new movie fragment
