+++ app/src/main/res/layout/movie_info_image_list_item.xml
@@ -7,11 +7,8 @@
     android:layout_gravity="center"
     android:padding="0dp"
     card_view:cardCornerRadius="0dp"
-    card_view:cardElevation="@dimen/movie_info_images_list_item_margin"
-    android:layout_marginRight="@dimen/movie_info_images_list_item_margin"
-    android:layout_marginEnd="@dimen/movie_info_images_list_item_margin"
-    android:layout_marginLeft="@dimen/movie_info_images_list_item_margin"
-    android:layout_marginStart="@dimen/movie_info_images_list_item_margin"
+    card_view:cardElevation="@dimen/movie_info_images_list_item_elevation"
+    android:layout_margin="@dimen/movie_info_images_list_item_margin"
     >
 
         <ImageView
