+++ app/src/main/res/values/dimens.xml
@@ -26,7 +26,8 @@
     <dimen name="vote_average_text_size">23sp</dimen>
     <dimen name="toolbar_shadow_height">7dp</dimen>
     <dimen name="expand_button_width">40dp</dimen>
-    <dimen name="movie_info_images_list_item_margin">3dp</dimen>
+    <dimen name="movie_info_images_list_item_margin">8dp</dimen>
+    <dimen name="movie_info_images_list_item_elevation">4dp</dimen>
     <dimen name="movie_info_photos_btn_size">28dp</dimen>
     <dimen name="movie_info_photos_btn_margin">5dp</dimen>
     <dimen name="movie_info_content_margin">32dp</dimen>
